Update
seeded database

+++ RecipeAPI/Models/RecipeCategoryItem.cs
@@ -2,7 +2,7 @@
 {
     public class RecipeCategoryItem
     {
-        public long RecipeId { get; set; }
+        public int RecipeId { get; set; }
         public int CategoryId { get; set; }
         public RecipeItem Recipe { get; set; } = null!;
         public CategoryItem Category { get; set; } = null!;
