+++ RecipeAPI/Models/RecipeIngredientItem.cs
@@ -2,8 +2,8 @@
 {
     public class RecipeIngredientItem
     {
-        public long RecipeId { get; set; }
-        public long IngredientId { get; set; }
+        public int RecipeId { get; set; }
+        public int IngredientId { get; set; }
         public RecipeItem Recipe { get; set; } = null!;
         public IngredientItem Ingredient { get; set; } = null!;
         public int? Amount { get; set; }
